requirements.txt
requests>=2.32.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
google-api-python-client>=2.126.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0